chore: 기존 다짐메시지 관련 파일 정리
