+++ src/pages/members/complete.tsx
@@ -8,11 +8,10 @@ import { useGetMemberProfileOfMe } from '@/api/endpoint_LEGACY/hooks';
 import AuthRequired from '@/components/auth/AuthRequired';
 import Responsive from '@/components/common/Responsive';
 import Text from '@/components/common/Text';
-import { LoggingClick } from '@/components/eventLogger/components/LoggingClick';
-import CardBack from '@/components/resolution/CardBack';
-import MemberCardOfMe from '@/components/resolution/MemberCardOfMe';
-import ResolutionModal from '@/components/resolution/ResolutionModal';
-import { useOpenResolutionModal } from '@/components/resolution/useOpenResolutionModal';
+import MemberCardOfMe from '@/components/members/upload/complete/MemberCardOfMe';
+import CardBack from '@/components/resolution/memberUploadComplete/CardBack';
+import ResolutionModal from '@/components/resolution/submit/ResolutionModal';
+import { useOpenResolutionModal } from '@/components/resolution/submit/useOpenResolutionModal';
 import { LATEST_GENERATION } from '@/constants/generation';
 import { MOBILE_MEDIA_QUERY } from '@/styles/mediaQuery';
 /**
@@ -33,8 +32,7 @@ const CompletePage: FC = () => {
       isActive: activity.generation === LATEST_GENERATION,
     }));
 
-  const { isOpenResolutionModal, onCloseResolutionModal, handleResolutionModalOpen, profileImage } =
-    useOpenResolutionModal();
+  const { isOpenResolutionModal, onCloseResolutionModal, profileImage } = useOpenResolutionModal();
 
   return (
     <AuthRequired>
